pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clawblink"
version = "0.3.0"
description = "No-Code AI Agent Builder from Chat. Describe an agent in plain English, ClawBlink builds and runs it."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "requests>=2.28.0",
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["builder*", "engine*", "providers*", "interfaces*"]
